@@ -1,27 +1,30 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.13'
+gem 'rails', '4.0.2'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
 
 gem 'pg'
-gem 'devise'
-gem 'carrierwave'
-gem 'asset_sync'
-gem 'fog'
+gem 'devise', '~> 3.2.2'
+gem 'carrierwave', '~> 0.9.0'
+gem 'asset_sync', '~> 0.5.4'
+gem 'fog', '~>1.18.0'
 gem 'rmagick'
-gem 'friendly_id', '~> 4.0.10'
-gem 'squeel'
+gem 'friendly_id', '~> 5.0.2'
+gem 'squeel', '1.1.1'
 gem 'coveralls', require: false
 gem 'faker'
+gem 'gemnasium'
+gem 'unf'
+gem 'protected_attributes'
 
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails'
+  gem 'coffee-rails'
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   gem 'therubyracer', :platforms => :ruby
@@ -29,10 +32,10 @@ group :assets do
   gem 'uglifier', '>= 1.0.3'
 end
 
-gem 'jquery-rails'
+gem 'jquery-rails', '~> 3.0.4'
 
 # To use ActiveModel has_secure_password
-gem 'bcrypt-ruby', '~> 3.0.0'
+gem 'bcrypt-ruby', '~> 3.1.2'
 
 # To use Jbuilder templates for JSON
 # gem 'jbuilder'
