@@ -1,5 +1,6 @@
 Kodiak::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
+  config.eager_load = false
 
   # The test environment is used exclusively to run your application's
   # test suite. You never need to work with it otherwise. Remember that
@@ -10,9 +11,6 @@ Kodiak::Application.configure do
   # Configure static asset server for tests with Cache-Control for performance
   config.serve_static_assets = true
   config.static_cache_control = "public, max-age=3600"
-
-  # Log error messages when you accidentally call methods on nil
-  config.whiny_nils = true
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
